feat: video lead asset (web) (#780)
* Feature: video lead asset in web article
* Updates package.json to sync brightcove versions
* Bumped major version number of BrightcoveVideo
