+++ packages/article/article.web.js
@@ -1,5 +1,4 @@
 import React, { Fragment } from "react";
-import { View } from "react-native";
 import PropTypes from "prop-types";
 import Ad, { AdComposer } from "@times-components/ad";
 
@@ -9,7 +8,7 @@ import { articlePropTypes, articleDefaultProps } from "./article-proptype";
 import ArticleHeader from "./article-header/article-header";
 import ArticleMeta from "./article-meta/article-meta";
 import ArticleBody from "./article-body/article-body";
-import LeadAssetComponent from "./article-lead-asset.web";
+import LeadAssetComponent from "./article-lead-asset";
 import articleTrackingContext from "./article-tracking-context";
 import getLeadAsset from "./get-lead-asset";
 import Topics from "./topics";
@@ -19,6 +18,7 @@ import {
   MainContainer,
   HeaderContainer,
   MetaContainer,
+  LeadAssetContainer,
   BodyContainer,
   HeaderAdContainer
 } from "./styles/responsive";
@@ -61,7 +61,6 @@ class ArticlePage extends React.Component {
           <Ad pos="header" style={adStyle} section={section} contextUrl={url} />
         </HeaderAdContainer>
         <MainContainer>
-          <LeadAssetComponent device="MOBILE" {...leadAssetProps} />
           <HeaderContainer>
             <ArticleHeader
               headline={headline}
@@ -71,24 +70,20 @@ class ArticlePage extends React.Component {
               isVideo={leadAssetProps.isVideo}
             />
           </HeaderContainer>
-          <View>
-            <MetaContainer>
-              <ArticleMeta
-                byline={byline}
-                publishedTime={publishedTime}
-                publicationName={publicationName}
-              />
-              <Topics topics={topics} device="DESKTOP" />
-            </MetaContainer>
-            <LeadAssetComponent device="DESKTOP" {...leadAssetProps} />
-            <BodyContainer>
-              <ArticleBody
-                content={content}
-                section={section}
-                contextUrl={url}
-              />
-            </BodyContainer>
-          </View>
+          <MetaContainer>
+            <ArticleMeta
+              byline={byline}
+              publishedTime={publishedTime}
+              publicationName={publicationName}
+            />
+            <Topics topics={topics} device="DESKTOP" />
+          </MetaContainer>
+          <LeadAssetContainer>
+            <LeadAssetComponent {...leadAssetProps} />
+          </LeadAssetContainer>
+          <BodyContainer>
+            <ArticleBody content={content} section={section} contextUrl={url} />
+          </BodyContainer>
         </MainContainer>
         <Topics topics={topics} />
         {displayRelatedArticles}
