+++ packages/article/fixtures/article-with-video-asset.json
@@ -99,11 +99,17 @@
       ],
       "leadAsset": {
         "type": "Video",
+        "brightcoveVideoId": "4084164751001",
+        "brightcovePolicyKey":
+          "BCpkADawqM0NK0Rq8n6sEQyWykemrqeSmIQqqVt3XBrdpl8TYlvqN3hwKphBJRnkPgx6WAbozCW_VgTOBCNf1AQRh8KnmXSXfveQalRc5-pyNlSod5XzP99If2U",
+        "brightcoveAccountId": "57838016001",
+        "paidonly": "false",
+        "caption": "This is video caption",
         "posterImage": {
           "id": "2d09f6bb-9106-4f8b-d4bc-f371672bd460",
           "title": "",
           "credits": "",
-          "caption": "",
+          "caption": "CCTV captures moment of the explosion",
           "crop": {
             "ratio": "16:9",
             "url":
