+++ packages/article/styles/article-header/responsive.web.js
@@ -2,8 +2,10 @@ import { View, Text } from "react-native";
 import withResponsiveStyles from "@times-components/responsive-styles";
 import { colours, fonts, fontSizes } from "@times-components/styleguide";
 
-export const HeadlineContainer = withResponsiveStyles(Text, {
-  base: () => `
+export const HeadlineContainer = withResponsiveStyles(
+  Text,
+  {
+    base: () => `
     font-size: ${fontSizes.headline}px;
     color: ${colours.functional.brandColour};
     margin-bottom: 8px;
@@ -11,12 +13,18 @@ export const HeadlineContainer = withResponsiveStyles(Text, {
     font-weight: 400;
     line-height: 30px;
   `,
-  mediumUp: () => `
+    mediumUp: () => `
     font-size: ${fontSizes.articleHeadline}px;
     line-height: 45px;
   `
-});
+  },
+  "HeadlineContainer"
+);
 
-export const LabelContainer = withResponsiveStyles(View, {
-  wideUp: () => "margin-top: 0px;"
-});
+export const LabelContainer = withResponsiveStyles(
+  View,
+  {
+    wideUp: () => "margin-top: 0px;"
+  },
+  "LabelContainer"
+);
