+++ packages/article/topics/styles/responsive.web.js
@@ -2,13 +2,21 @@ import { View } from "react-native";
 import withResponsiveStyles from "@times-components/responsive-styles";
 import config from "../../styles/responsive-config";
 
-export const TopicsContainer = withResponsiveStyles(View, {
-  base: () => config.articleContainerPadding,
-  mediumUp: () => config.mediumBpPositioning,
-  wideUp: () => `display: none`
-});
+export const TopicsContainer = withResponsiveStyles(
+  View,
+  {
+    base: () => config.articleContainerPadding,
+    mediumUp: () => config.mediumBpPositioning,
+    wideUp: () => `display: none`
+  },
+  "TopicsContainer"
+);
 
-export const TopicsMetaContainer = withResponsiveStyles(View, {
-  base: () => `display: none;`,
-  wideUp: () => `display: block;`
-});
+export const TopicsMetaContainer = withResponsiveStyles(
+  View,
+  {
+    base: () => `display: none;`,
+    wideUp: () => `display: block;`
+  },
+  "TopicsMetaContainer"
+);
