+++ packages/brightcove-video/brightcove-player.web.js
@@ -223,7 +223,7 @@ class BrightcoveVideo extends Component {
     /* eslint jsx-a11y/media-has-caption: "off" */
     // Added a wrapping div as brightcove adds siblings to the video tag
     return (
-      <div>
+      <div style={{ width: this.props.width, height: this.props.height }}>
         <video
           id={this.state.id}
           style={{ width: this.props.width, height: this.props.height }}
