+++ packages/gradient/package.json
@@ -44,7 +44,7 @@
   "homepage": "https://github.com/newsuk/times-components#readme",
   "devDependencies": {
     "@storybook/react-native": "3.3.15",
-    "@times-components/brightcove-video": "1.12.1",
+    "@times-components/brightcove-video": "2.0.0",
     "@times-components/eslint-config-thetimes": "0.1.4",
     "@times-components/jest-configurator": "0.2.2",
     "dextrose": "1.7.3",
