+++ packages/image/image-prop-types.js
@@ -5,11 +5,12 @@ const { style: ViewPropTypesStyle } = ViewPropTypes;
 
 export const defaultProps = {
   style: {},
-  uri: ""
+  uri: "",
+  aspectRatio: undefined
 };
 
 export const propTypes = {
   uri: PropTypes.string,
-  aspectRatio: PropTypes.number.isRequired,
+  aspectRatio: PropTypes.number,
   style: ViewPropTypesStyle
 };
